refactor rest/client pkg
