+++ e2e/hook/host_client.go
@@ -5,41 +5,14 @@ import (
 	"fmt"
 
 	"github.com/cucumber/godog"
-	"k8s.io/apimachinery/pkg/runtime"
-	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
-	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
-	"k8s.io/client-go/tools/clientcmd"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	"github.com/konflux-workspaces/workspaces/e2e/hook/internal"
 	"github.com/konflux-workspaces/workspaces/e2e/pkg/cli"
 	tcontext "github.com/konflux-workspaces/workspaces/e2e/pkg/context"
-
-	toolchainv1alpha1 "github.com/codeready-toolchain/api/api/v1alpha1"
-	workspacesiov1alpha1 "github.com/konflux-workspaces/workspaces/operator/api/v1alpha1"
-	restworkspacesiov1alpha1 "github.com/konflux-workspaces/workspaces/server/api/v1alpha1"
+	wrest "github.com/konflux-workspaces/workspaces/e2e/pkg/rest"
 )
 
 func injectHostClient(ctx context.Context, sc *godog.Scenario) (context.Context, error) {
-	apiConfig, err := clientcmd.NewDefaultClientConfigLoadingRules().Load()
-	if err != nil {
-		return nil, fmt.Errorf("error building config: %v", err)
-	}
-
-	cfg, err := clientcmd.NewDefaultClientConfig(*apiConfig, &clientcmd.ConfigOverrides{}).ClientConfig()
-	if err != nil {
-		return nil, fmt.Errorf("error building config: %v", err)
-	}
-
-	internal.MutateConfig(cfg)
-
-	scheme := runtime.NewScheme()
-	utilruntime.Must(clientgoscheme.AddToScheme(scheme))
-	utilruntime.Must(workspacesiov1alpha1.AddToScheme(scheme))
-	utilruntime.Must(restworkspacesiov1alpha1.AddToScheme(scheme))
-	utilruntime.Must(toolchainv1alpha1.AddToScheme(scheme))
-
-	c, err := client.New(cfg, client.Options{Scheme: scheme})
+	c, err := wrest.BuildDefaultHostClient()
 	if err != nil {
 		return nil, fmt.Errorf("error building client: %v", err)
 	}
