+++ e2e/pkg/rest/config.go
@@ -1,4 +1,4 @@
-package internal
+package rest
 
 import (
 	"os"
@@ -22,11 +22,11 @@ var mutationFuncs = []configMutationFunc{
 	setTLSFromEnv,
 }
 
-// MutateConfig adjusts a pre-populated rest.Config to adapt to the
+// mutateConfig adjusts a pre-populated rest.Config to adapt to the
 // configuration we've been given from the environment.  For now, this only
 // disables TLS validation if the environment variable `E2E_USE_INSECURE_TLS`
 // is set.
-func MutateConfig(cfg *rest.Config) {
+func mutateConfig(cfg *rest.Config) {
 	for _, f := range mutationFuncs {
 		f(cfg)
 	}
